indexer/requirements.txt
watchdog>=4.0.0
qdrant-client>=1.9.0
httpx>=0.27.0
python-frontmatter>=1.1.0
tiktoken>=0.7.0
